+++ apps/invitees/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+class InviteesConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'apps.invitees'